backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-environ
django-auditlog
drf-spectacular
gunicorn
whitenoise==6.6.0
